feat: adds try/catch to all utils files
