+++ src/utils/formatZipcode.ts
@@ -1,5 +1,9 @@
 export default function formatZipcode(zipcode: string): string {
-	if (!zipcode) return "not_found";
+	try {
+		if (!zipcode) return "not_found";
 
-	return zipcode.replace(/\D/g, "").slice(0, 8);
+		return zipcode.replace(/\D/g, "").slice(0, 8);
+	} catch {
+		return "not_found";
+	}
 }
