+++ src/utils/index.ts
@@ -5,3 +5,6 @@ export { default as isDateValid } from "./isDateValid";
 export { default as normalizeCity } from "./normalizeCity";
 export { default as formatZipcode } from "./formatZipcode";
 export { default as isAdult } from "./isAdult";
+export { default as capitalizeFirst } from "./capitalizeFirst";
+export { default as getFirstName } from "./getFirstName";
+export { default as formatRegisterFormValues } from "./formatRegisterFormValues";
