+++ src/utils/normalizeCity.ts
@@ -1,11 +1,15 @@
 export default function normalizeCity(city: string): string {
-	if (city === "not_found" || !city) return "not_found";
+	try {
+		if (city === "not_found" || !city) return "not_found";
 
-	return city
-		.normalize("NFD")
-		.replace(/[\u0300-\u036f]/g, "")
-		.replace(/['"]/g, "")
-		.replace(/\([^()]*\)/g, "")
-		.toUpperCase()
-		.trim();
+		return city
+			.normalize("NFD")
+			.replace(/[\u0300-\u036f]/g, "")
+			.replace(/['"]/g, "")
+			.replace(/\([^()]*\)/g, "")
+			.toUpperCase()
+			.trim();
+	} catch {
+		return "not_found";
+	}
 }
