+++ src/utils/stringifyBigInt.ts
@@ -1,14 +1,18 @@
 export default function stringfyBigInt(obj: Record<string, unknown>) {
-	return Object.keys(obj).reduce((previousValue, currentKey) => {
-		let currentObjValue = obj[currentKey];
+	try {
+		return Object.keys(obj).reduce((previousValue, currentKey) => {
+			let currentObjValue = obj[currentKey];
 
-		if (typeof currentObjValue === "bigint") {
-			currentObjValue = currentObjValue.toString();
-		}
+			if (typeof currentObjValue === "bigint") {
+				currentObjValue = currentObjValue.toString();
+			}
 
-		return {
-			...previousValue,
-			[currentKey]: currentObjValue,
-		};
-	}, {});
+			return {
+				...previousValue,
+				[currentKey]: currentObjValue,
+			};
+		}, {});
+	} catch {
+		return obj;
+	}
 }
